Rewrote Timer.
Timer:  Storing functors (to members) instead of ptrs to TimerActions. Got
rid of start actions because they can be called by the owner when the timer
is started.
TimerAction: Made a "tag" class instead. Took out pure virtual functions to
be overridden and replaced with a pure virtual dtor that is implemented. This
class cannot be instantiated and subclasses don't need to override the pure
virtual dtor.


svn path=/trunk/; revision=55