Undid the rollback. Using boost::function for Timer.
In revision 57, I rolled back to 55. Then I experimented with boost::function
and found that templates might pose a solution for the Timer class. They work
for the most part, so rolled back revision 58 to 56. Still haven't made use of
the timer class and ironed out the details. Commented out places in 
MainGameScene where Timer class was used.


svn path=/trunk/; revision=59